Update .classpath
Most of the .classpath was deleted in a recent commit. This is what was causing errors with gradle, the package, and include statements in files. It did not, in fact cause the getter/setter generating error on my machine - that error was because you must left click somewhere in the code you are editing before generating getters and setters. However, the getter/setter error on my machine was different from the error seen by the rest of the team, which may have been caused by one of the two issues above or something entirely different.